CONSEGUI

+++ include/generators.hpp
@@ -8,4 +8,4 @@ ParsedModel generateSphere(float radius, int slices, int stacks);
 ParsedModel generateBox(float dimension, int divisions);
 ParsedModel generateCone(float radius, float height, int slices, int stacks);
 ParsedModel generateCylinder(float radius, float height, int sides);
-ParsedModel generatePatch(std::string &patchfile, float tesselationLevel);
+ParsedModel generatePatch(std::string &patchfile, int tesselationLevel);
